requirements.txt
Django==3.1.5
python-decouple==3.4
psycopg2-binary==2.8.6
model-bakery==1.2.1
pillow==8.1.0
django-model-utils==4.1.1
django-reverse-admin==2.9.4
